django/django_webadmin/requirements.txt
Django == 1.9
django_mptt == 0.8.3
djangorestframework == 3.3.2
